requirements.txt
scikit-learn==0.22.2.post1
pandas==1.0.3
Django==3.0.5
PyYAML==5.3.1
requests